feat: allow overriding default task deadline
This has come up with the work on our Translations model training pipeline. Full training runs are expected to take weeks or even a month to complete. Easily allowing the entire graph to have a longer deadline is the easiest way to cope with this, and I imagine other things could benefit as well, eg: fuzzing.