feat: allow overriding default task deadline (#290)
* feat: allow overriding default task deadline

This has come up with the work on our Translations model training pipeline. Full training runs are expected to take weeks or even a month to complete. Easily allowing the entire graph to have a longer deadline is the easiest way to cope with this, and I imagine other things could benefit as well, eg: fuzzing.

* style: pre-commit.ci auto fixes [...]

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>