feat: transform that supports generic chunking of tasks
Chunking is a common feature in taskgraphs to run jobs in parallel.

The `chunk` sequence is used to split a job into `total-chunks` chunks, and substitute `this_chunk` and `total_chunks` into any of the `substitution-fields`. Typically this is needed in `label` and `description` as well as at least one field that impacts the running job (eg: `env` or `command`).