fix: relax PyYaml specification to work around 'mach vendor' issue in Gecko
I don't understand the details, but somehow the recent Cython 3.0
release broke the ability to build PyYaml. There's currently no fix
other than to downgrade to PyYaml 5.3.1.

Unfortunately, Taskgraph is causing a conflict as it has PyYaml >=
5.4.0. Tracing back through history, there wasn't any particular reason
for this, other than me wanting to be able to make some assertions
about our dependencies and taking advantage of the fact that Taskgraph
was a new package.

The Gecko bug is:
https://bugzilla.mozilla.org/show_bug.cgi?id=1843902

The upstream PyYaml issue is:
https://github.com/yaml/pyyaml/issues/601