fix: from-deps: don't override all existing dependencies; don't add fetch entry for kinds without fetches
The former is arguably more than a "fix" in that it changes `from-deps` in a way that it is no longer the sole thing that manages upstream dependencies. In my opinion, if we are considering `from-deps` the official way to collect results from chunked/parallelized tasks, this is a necessary change, as we cannot predict the other things that such tasks may also depend on.

The latter is a pure bug fix: we shouldn't add a `fetch` block for kinds that have no `fetches` listed in `from-deps`.

(I can separate these two if it's useful.)