fix: ensure we can resolve default Parameters even if no repo is detected
This is being implemented to support a Gecko feature where they want to
be able to run `./mach bootstrap` from a non-repo extracted artifact.
Currently, since bootstrap depends on Taskgraph, it is not possible.

AIUI, they don't need to actually generate a graph and only the default
parameters need to be able to work without a repo.

Bug: 1847288