fix: bypass Git's safe directory feature in 'run-task'
As of Git 2.35.2, a security feature to mitigate a CVE on multi-user
setups:
https://github.blog/2022-04-12-git-security-vulnerability-announced/#cve-2022-24765

Basically Git will error out if the checkout dir isn't owned by the
current user. This impacts `run-task` because checkouts can come from
caches meaning they were cloned by a different task user.

As long as we aren't sharing caches across levels or pools (which we
aren't), that other user is just as trusted as the current one. So it's
safe to disable this check.

Issue: #345