refactor!: remove unused util.decision.make_decision_task
This was added in 2eec86155dc2e46e90da6930664df5b39ef24a3a. It appears
to have been unused then, and is unused now. I would guess this used to
be used in Gecko, so made its way here, but then got dropped in Gecko at
some point.

This is technically backwards incompatible, so let's land with the other
major changes.