fix(init): Don't set 'base_ref' on github-release events
We were previously setting this to the tag of the release, but run-task
doesn't handle the case where base_ref is a tag well. This could be
fixed, but the release tag is the `head_ref` rather than the `base_ref`.
So it's best to just not pass it in at all in this case.