feat: use new batched queries in task replacement
Batching queries to get from task index to status reduces the number of
(sometimes trans-continental) queries from 2*900+ to ~2. This reduces the
time spent replacing tasks from 20% to 75% depending on the use-case.

The 20% improvement in wall time was observed when running
`mach taskgraph morphed` in a CI worker, while the 75% improvement
was observed in a developer machine in France running `mach taskgraph full`.

More information in https://github.com/taskcluster/taskcluster-rfcs/pull/189.