ci: add 'docker-image' task for 'image-builder'
It might be a little odd to use the `image_builder` image to build the
`image_builder` image. But doing this should be fine and is analogous to
compilers using earlier versions of themselves to compile themselves.

Besides, if we ever break the image we'd just be back to where we are
today, needing to build it manually.

Note for now I'm simply focusing on adding builds and not worrying about
any release mechanisms. A later PR will add a `push-image` task and
somehow figure out how to automatically bootstrap to this.