fix!: stop merging "by-*" keys in `util.templates.merge`
BREAKING CHANGE: This can impact how `task-defaults` gets merged into
each task definition.

The default merge behaviour would mix the `by-<key>` key in with
non-keyed-by configs. This is a bug and certainly not what the user
would be expecting.

It is possible that the user intended to mix two `by-<key>` configs
together, but this case should be extremely rare.

This is a port of the following `gecko_taskgraph` patch:
https://hg.mozilla.org/mozilla-central/rev/0fe5ffc10ff8488d780c5c3cbcff1b43ca974cf7