feat: add a separate log for optimizations (fixes #523) (#525)
* feat: use separate python logger for optimizations

This is a precursor that will allow us to log these more verbosely and to a different location than the rest of the taskgraph log messages.

* fix: IndexSearch tests to pass for the right reasons

Without this, we ended up passing due to entering the KeyError block in IndexSearch rather than the actual blocks we were supposed to.

* test: add tests for optimization log messages

* feat: always write optimization log information verbosely to a separate file

This is a _little_ bit nasty because we don't have a great log configuration story in taskgraph to begin with, but it works. The `taskgraph_decision` parts are also untested because we don't have a good testing story there. I don't particularly want to shave those yaks as part of this...