fix: get_ancestors should ignore 404s (#569)
This can happen when attempting to fetch tasks (either directly or indirectly) that have expired. My view on this is that such tasks should be treated as no longer being dependencies since they no longer exist.

I can see a case for maybe making this optional behaviour, but I don't know of any use cases where we'd want to raise an exception for this case. Certainly when being used as part of actions or other CI code it ends up being a ticking time bomb.