feat: add 'project' and 'trust-domain' to set of tags applied to all tasks
Task tags will show up in the Firefox-CI ETL, so adding these tags will
make it possible to group costs by project and trust domain
respectively.

Note that Taskgraph is already inconsistent with the global tags it
applies. There is `worker-implementation` and `createdForUser`. I
decided to use dashes instead of camel case, but don't have a strong
preference. We should make the tag format consistent, but that's a
breaking change which I'll leave for another time.