refactor!: strip 'build-' prefix from docker image task labels
BREAKING CHANGE: `build-` prefix stripped from docker image tasks

This is being done to facilitate merging some docker related logic with
gecko_taskgraph. The issue with adding a `build-` prefix to the labels,
is that it is common for there to exist a `build` kind.

While it isn't best practice to parse labels to obtain information like
the kind, in reality this happens pretty frequently. So making this
change:

1. Makes the label consistent with Gecko (needed to merge some docker
   related logic)
2. Makes the label consistent with other tasks (prefix should just be
   the kind name)
3. Reduces footguns for folks that are parsing the label.