actions: look up action tasks by revision in fetch_graph_and_labels (#589)
Projects on github don't have a meaningful pushlog-id, so we don't use
it in the index path for action tasks.  When building the
label-to-taskid mapping for a revision, look up action tasks in the
index using the head_rev, where we can actually find them.

This can potentially do the wrong thing if a revision is pushed to two
different branches, where we'd be mixing up the graphs, but we don't
currently include the branch name in the index paths.