fix!: only add target_tasks_method filter if no other filters were specified
BREAKING CHANGE: The target_tasks_method is now only implicitly added if
no other filters exist.

This is needed to fix a bug in Gecko's `./mach try` infrastructure,
however the change makes sense on its own. Currently there's no way to
forego target_tasks_method, and this allows us to do that.

Furthermore, there are no other built-in filters, so it's unlikely that
this will impact any existing consumers.