Hash the uncompressed contents of docker context tar
With an increasing number of docker images, the overhead of hashing the
compressed contents of the docker context tars becomes visible when
loading the toolchain kind during Firefox builds. But the really
interesting part for the hash is the uncompressed contents. Whether they
are compressed with gzip at level 9, gzip level 1, bz2 or zstd, it
doesn't matter for the resulting docker image. So it makes sense to
only hash the uncompressed contents, which is much faster.