Back out "feat(caches): support opting into or out of specific name caches"
This backs out commit a70a57c72c794d992075548eefc03755e387c9f1.

For context I should have held off landing this as I ended up deciding
that it was better to have well defined caches that need to be opted
into explicitly.

I want to avoid having a release with this regex feature to ensure we
don't have users accidentally using it and then needing to stop using it
on the next release.