refactor!: revert {task_workdir} interpolation in run-task environment
In hindsight, this ended up being a bad idea and caused lots of subtle
issues in Gecko. As an alternative, I landed a change to get
generic-worker itself to set $TASK_WORKDIR.

This will take awhile to propagate to all pools, but in the meantime
it's cleanest to revert back to the abspath method.