Add a simple test for `taskgraph action`
To be able to test without relying on actions that are always generated
too much, I monkeypatch `registry._load` so it simply doesn't import the
modules registering those actions.

Adding that test found that parameters should be parsed with
`strict=False` to match other CLI actions. The return value isn't
strictly necessary but I added it to be consistent with other functions
in the CLI.