Add a workflow to run auto updates for pre-commit
This replaces the pre-commit.ci app's autoupdate.

The one caveat with this is that it doesn't check for PRs that are
already opened on that branch. I did that to keep it as simple as
possible. It's not a real issue as github will detect that and close the
previous PR and then open a new one but it might be a bit spammy if we
don't merge those PRs