fix: fallback to guessing repo root in config.py
We made the repo root detection smarter by using util/vcs.py to search
for the `.git` directory. However, this broke a MacOS build task in
Gecko CI because no `.git` directory exists. I'm not entirely sure why
that happens, as the Taskgraph command still needs to run from the repo
root.

Ay any rate, this patch re-adds the old logic as a fallback in the event
util/vcs.py wasn't able to detect the repo.