Update taskcluster to 91+ to fix get_artifact
This fixes a regression from 384afee759351f22ab9e0a43ac066f3fa5f3b6d9
where we started using the taskcluster python client instead of making
requests manually.

The artifacts route is special because the API returns some JSON as part
of a 303 to the artifact's content URL. The previous code followed that
303. The python client before tascluster 91 does not. So instead we're
getting the body which looks like this: `{"type": "s3/...", "url":
"whereverthe303pointsat"}`.

I also fixed the test as the previous one was basically ignoring the
fact that taskcluster was returning a 303 and was mocking as if it
didn't exist.

Fixes #812