Force `Graph` to actually use frozensets
Because type annotations in python are a "I wish people would respect my
authority" kind of feature, we cannot guarantee that a `Graph` is
actually created with frozensets as attributes. That is the case all
throughout taskgraph where we're passing sets. By forcing those
attributes to be frozensets we can ensure that `Graph` is hashable which
will allow us to use `functools.lru_cache` to cache some expensive
functions in here.

Unfortunately, there's no easy way of transforming the inputs passed
into a dataclass, since in both `__init__` and `__post_init__`, the
structure is already frozen. By using a dataclass as the parent of
`Graph` instead of `Graph` itself, we can work around the issue while
keeping the frozenness of the dataclass post init without having to
resort to working around dataclass when transforming the attributes