fix: avoid proxy by default in get_index_url
This aligns with get_artifact_url. Typically if you're building a url
rather than actually fetching something, then it's because the
downloading will happen in a downstream task, where the proxy might not
be available, or might have a different url. So defaulting to not using
it makes the most sense.