feat: add support for outputting a mermaid flowchart of the kind graph (#834)
Especially in repositories with large graphs, it can be useful to visualize the relationship between the various kinds.

I looked at directly outputting svg files of the graphs here, but decided against it because rendering mermaid diagrams requires a browser or puppeteer AFAICT; and I don't want to add that dependency.

I also looked at adding support for visualizing entire graphs, but those quickly get very unwieldy and I decided against it. (This might be useful in conjunction with a new `--target-task` option, but that's more than I'm willing to take on at this time.)