feat: run tests against 3.14 free-threaded python (#837)
I'm looking to start experimenting with threads for parallel kind generation. To make this worthwhile, we need to support free-threaded python. It looks like tests already pass here, so this should be a no-brainer for now. I've done some basic usage testing as well, and everything seems to work (even with free-threaded python...which makes sense seeing as we don't use threads anywhere at the moment).