Register the optimization schema with `register_strategy`
Instead of hardcoding the schema, make it part of the registration of
the strategy, this allows for people to add their own strategy without
having to work around `OptimizationSchema` which is painful and brittle.

The slight drawback from this is that we lose the documentation in
the task schema about what is possible since it's now deferred to the
strategies themselves, and that we now validate optimizations twice,
first to check that it's a dict then to check that it matches the
strategy.

Fixes #368